git-svn-id: svn://hevc.kw.bbc.co.uk/svn/jctvc-tmuc/trunk@1670 ce97f6c2-8f86-0410-b44d-af84b13ec34f

+++ source/Lib/TLibEncoder/NALwrite.cpp
@@ -132,7 +132,7 @@ void write(ostream& out, const OutputNALUnit& nalu)
     } while (true);
 
 #if TILES_DECODER
-    UInt uiDistance = found - rbsp.begin();
+    UInt uiDistance = (UInt)(found - rbsp.begin());
     for ( UInt uiMrkrIdx = 0; uiMrkrIdx < uiTileMarkerCount ; uiMrkrIdx++ )
     {    
       UInt uiByteLocation   = nalu.m_Bitstream.getTileMarkerLocation( uiMrkrIdx );
